(refactor) Destructure error and isLoading variables from SWR hooks
This PR refactors some SWR-based hooks to correctly destructure the `error` and `isLoading` variables directly from the hooks.
These hooks were previously setting an `isError` variable, which was incorrect as the `error` property returned by SWR is an Error object, not a boolean.
It also fixes some instances where the `isLoading` variable was not being used.

It also restores husky, which should fix the pre-commit and pre-push hooks.