(fix) Disable number input scroll wheel functionality (#226)
This PR disables the scroll wheel functionality in [NumberInput](https://react.carbondesignsystem.com/?path=/docs/components-numberinput--overview) fields in the Stock Management app by setting the `disableWheel` prop. Before this change, when a NumberInput field was focused and the user scrolled past with their mouse wheel, the value in the input would increment or decrement. This can lead to accidental value changes that the user might overlook. Setting the prop fixes makes the NumberInput ignore mouse wheel events when focused, reducing the likelihood of unintentional value changes.