(chore) Add bundle size reporting to GitHub Actions (#228)
This PR adds a GitHub Actions workflow to report the bundle size of the library. It uses
`compressed-size-action` to compute the bundle size and `turborepo-gh-artifacts` to
serve as a local cache server for Turborepo. With this setup, we can be able to track bundle
size changes and flag any increase in bundle size.